Support file chooser on macOS
