+++ webextensions/native-messaging-host/host.go
@@ -10,8 +10,6 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"github.com/harry1453/go-common-file-dialog/cfd"
-	"github.com/harry1453/go-common-file-dialog/cfdutil"
 	rotatelogs "github.com/lestrrat/go-file-rotatelogs"
 	"github.com/lhside/chrome-go"
 	"io/ioutil"
@@ -216,29 +214,6 @@ func ChooseFileAndRespond(params RequestParams) {
 	if err != nil {
 		log.Fatal(err)
 	}
-}
-
-func ChooseFile(params RequestParams) (path string, errorMessage string) {
-	result, err := cfdutil.ShowOpenFileDialog(cfd.DialogConfig{
-		Title: params.Title,
-		Role:  params.Role,
-		FileFilters: []cfd.FileFilter{
-			{
-				DisplayName: params.DisplayName,
-				Pattern:     params.Pattern,
-			},
-		},
-		SelectedFileFilterIndex: 0,
-		FileName:                params.FileName,
-		DefaultExtension:        params.DefaultExtension,
-	})
-	if err == cfd.ErrorCancelled {
-		return result, ""
-	} else if err != nil {
-		log.Fatal(err)
-		return "", err.Error()
-	}
-	return result, ""
 }
