+++ webextensions/native-messaging-host/host_darwin.go
@@ -12,12 +12,33 @@ package main
 import (
 	"encoding/json"
 	"github.com/lhside/chrome-go"
+	"github.com/ncruces/zenity"
 	"log"
 	"os"
 )
 
+
+func ChooseFile(params RequestParams) (path string, errorMessage string) {
+	filename, err := zenity.SelectFile(
+		zenity.Title(params.Title),
+		zenity.Filename(params.FileName),
+		zenity.FileFilters{
+			{params.DisplayName, []string{params.Pattern}, true},
+		})
+	if err == zenity.ErrCanceled {
+		return filename, ""
+	} else if err != nil {
+		log.Fatal(err)
+		return "", err.Error()
+	}
+	return filename, ""
+}
+
+
 func FetchOutlookGPOConfigsAndResponse() {
 	response := OutlookGPOConfigsResponse{}
+
+	// NOT IMPLEMENTED YET
 
 	body, err := json.Marshal(response)
 	if err != nil {
